rename loaders to loader
