+++ go.mod
@@ -1,6 +1,6 @@
 module fyne.io/x/fyne
 
-go 1.15
+go 1.16
 
 require (
 	fyne.io/fyne/v2 v2.0.2
