+++ widget/loader/three_dots.go
@@ -1,4 +1,4 @@
-package loaders
+package loader
 
 import (
 	"image/color"
